Task 3
You have been given the coupondunia database which has the following
tables that you need to use:
-Coupon
-Website
-CouponCategories
-CouponCategoriyInfo
-CouponSubCategories

You need to implement dynamic filters for these coupons. Create a 1
page UI which displays all 60 coupons on 1 webpage. You need to provide
3 types of filters over these coupons:

1. Filter by coupon type [Deal/Coupon]

2. Filter by Stores

3. Filter by Categories

All the data required for these filters is available in the database.

The filters should be dynamic. When someone wants to filter to show
only coupons, the calls must be made via ajax.

This page should also have a "Download Data" button. On click of this
button, the filtered data, shown on the page, must be written to an
Excel file.

You are expected to implement good architecture/design with classes for
data wherever needed.
I must stress that you need to be well-versed in OOPS concepts, PHP,
MySQL, AJAX, JavaScript and HTML/CSS.
We will review the code you write and then ask you to correct it if and
till we think it is using all concepts correctly.
As part of the review, we may ask you to go through certain parts of
the tutorial again (if we think it's required).